💡 마무리 주석 정리
