+++ src/main/java/com/multi/biz/PmediBiz.java
@@ -23,7 +23,7 @@ import com.multi.vo.PmediVo;
  *
  *	2022. 7. 25.		 najune				 get_medi 추가
  *
- *	2022. 8. 05.		 						 add 추가
+ *	2022. 8. 05.		 					  add 추가
  *
  * =========================================================
  */
