+++ src/main/java/com/multi/controller/APIController.java
@@ -14,6 +14,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import com.multi.biz.SlistBiz;
 import com.multi.biz.SmediBiz;
 import com.multi.frame.Util;
+import com.multi.restapi.DataAPI;
+import com.multi.restapi.OCRBoxAPI;
 import com.multi.restapi.OCREnvelopeAPI;
 import com.multi.vo.MfVo;
 import com.multi.vo.SlistVo;
@@ -23,7 +25,7 @@ import com.multi.vo.UsersVo;
 /**
  * @author noranbear
  * @date 2022. 8. 10.
- * @version 2.2
+ * @version 3.0
  * @description
  *
  *
@@ -33,6 +35,8 @@ import com.multi.vo.UsersVo;
  *  2022. 8. 10.		noranbear		  First Creation
  *
  *	2022. 8. 11.						   eocraddimpl 생성
+ *
+ *	2022. 8. 18. 						 api 관련 함수 다 옮기기
  *
  * =========================================================
  */
@@ -46,14 +50,145 @@ public class APIController {
 	@Autowired
 	SmediBiz smbiz;
 	
-	// 이미지 저장 경로
-	@Value("${userdir2}")
+	// ocrbox 이미지 저장 경로
+	@Value("${userdir}")
 	String userdir;
+		
+	// ocrenvelope 이미지 저장 경로
+	@Value("${userdir2}")
+	String userdir2;
+	
+  	@Autowired
+	OCRBoxAPI bapi;
+  	
+	@Autowired
+	DataAPI dapi;
 	
 	@Autowired
 	OCREnvelopeAPI eapi;
 		
+	
+	/**
+	 * 메인페이지에서 ocr창으로 들어온 1. 이미지 저장, 2. ocrbox search, 
+	 * 3. 해당 내용 DB(slist, smedi)에 저장하는 함수
+	 * @param mf 이미지 파일
+	 * @param session 현재 유저 id를 가져올 HttpSession
+	 */
+	@RequestMapping("/ocraddimpl")
+	public String ocraddimpl(Model m, MfVo mf, HttpSession session) {
+		int listId = 0;				// 스캔한 약곽 정보를 담은 slistVo의 id를 담음
+		String imgname = null;		// mf의 이미지 이름을 담음
+		String uid = null;			// users에서 받을 유저 id를 담음
+		String name = null;			// 스캔한 약곽 이름을 담음
 		
+		UsersVo users = null;		// HttpSession에서 가져오는 유저 정보를 담음
+		SlistVo slist = null;		// 스캔한 약곽 정보를 slist tbl에 넣을 때 사용
+		
+
+		// [1] 약 이미지 저장
+		// 1-1. 이미지 이름을 가져온다.
+		imgname = mf.getMf().getOriginalFilename();		// mf 앞에 fake path가 붙기 때문에 필요
+
+		// 조건 1: 이미지가 존재할 때 실행
+		if(!(imgname.isEmpty())) { 
+			
+			// [2] DB에 정보 저장
+			// 2-1. (비)회원 ID를 가져온다.
+			// 유저일 때
+	        if(session.getAttribute("signinusers") != null){	
+	        	users = (UsersVo) session.getAttribute("signinusers");		// HttpSession에서 UsersVo 가져옴
+	            uid = users.getId();
+	        
+	        // 비회원일 때
+	        }else {
+	        	uid = "none0";		// DB에 저장되어 있는 비회원 아이디
+	        }
+	        
+	        slist = new SlistVo(uid, imgname);  
+	         
+			try {
+				
+				// 2-2.스캔내역 tbl에 데이터 넣는다.
+				slibiz.registerbox(slist);
+				
+				// 1-2. 이미지를 해당 경로에 저장한다.
+				Util.saveFile(mf.getMf(), userdir);
+				
+				// [3] ocrbox 스캔
+				// 3-1. ocrbox로 스캔한다.
+				Object result = bapi.boxapi(imgname);
+				
+				// 3-2. 리턴된 Object에서 Json parsing해서 약이름만 빼낸다.
+				JSONObject jo = (JSONObject) JSONValue.parse(result.toString());
+				JSONArray images = (JSONArray) jo.get("images");
+				//System.out.println("1 : " + jo1);
+				JSONObject obj = (JSONObject) images.get(0);
+				//System.out.println("2 : " + obj);
+				JSONObject title = (JSONObject) obj.get("title");
+				//System.out.println("3 : " + obj2);
+				name = (String) title.get("inferText");
+				//System.out.println("4 : " + name);
+				
+				// 3-3. 구한 약이름을 화면으로 보낸다.
+				m.addAttribute("ocrresult", "ok");
+				m.addAttribute("resultname", name);
+				
+				// 2-3. 스캔약 tbl에 데이터를 넣는다.
+				slist = slibiz.gettheone(slist);	// smedi는 sid를 필요로 하기 때문에 DB에 있는 
+													// (id를 가지고 있는)slist를 다시 꺼내온다.
+
+				// slist tbl에 해당 정보가 존재할 때
+				if(slist.getId() != 0) {		// int id 값이 없는 경우 0 리턴
+					listId = slist.getId();
+				    SmediVo smedi = new SmediVo(name, listId);
+					smbiz.register(smedi);
+				}
+				
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+		}
+		// page breadcrumb
+		m.addAttribute("pagename", "검색");
+		m.addAttribute("pagename2", "검색");
+		
+		return "index";
+
+	}
+	
+	
+	/**
+	 * 약 디테일 페이지 연결
+	 * @return medidetail.html
+	 */
+	@RequestMapping("/medidetail")
+	public String mdetail(Model m, String item) {
+
+		Object obj = dapi.dataapi(item);
+		//System.out.println("result 값 : " + obj);
+       
+		// Object를 JSONObject으로 변환
+		JSONObject jo = (JSONObject) JSONValue.parse(obj.toString());
+       
+		// jo에서 JSONObject으로 body 뽑아내기
+		JSONObject jo1 = new JSONObject();
+		jo1 = (JSONObject) jo.get("body");
+       
+		// body에서 JSONArray로 items 뽑아내기
+		JSONArray ja = new JSONArray();
+		ja = (JSONArray) jo1.get("items");
+       
+		m.addAttribute("item", ja);
+       
+		// page breadcrumb
+		m.addAttribute("pagename", "약 상세정보");
+		m.addAttribute("pagename2", item);
+	
+		m.addAttribute("center", "medidetail");
+		return "index";
+	}
+	   
+	   
 	/**
 	 * 메인페이지에서 ocr창으로 들어온 
 	 * 1. 이미지 저장
@@ -103,7 +238,7 @@ public class APIController {
 				slibiz.registerbox(slist);
 				
 				// 1-2. 이미지를 해당 경로에 저장한다.
-				Util.saveFile(mf.getMf(), userdir);
+				Util.saveFile(mf.getMf(), userdir2);
 				
 				
 				// [3] ocr 스캔 후 병원명, 조제일자, 투약일수, 투약시간을 화면으로 전달
