+++ src/main/java/com/multi/controller/MainController.java
@@ -7,11 +7,7 @@ import java.util.List;
 
 import javax.servlet.http.HttpSession;
 
-import org.json.simple.JSONArray;
-import org.json.simple.JSONObject;
-import org.json.simple.JSONValue;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -24,11 +20,7 @@ import com.multi.biz.PmediBiz;
 import com.multi.biz.SlistBiz;
 import com.multi.biz.SmediBiz;
 import com.multi.biz.UsersBiz;
-import com.multi.frame.Util;
-import com.multi.restapi.DataAPI;
-import com.multi.restapi.OCRBoxAPI;
 import com.multi.vo.AlarmVo;
-import com.multi.vo.MfVo;
 import com.multi.vo.MymediVo;
 import com.multi.vo.PlistVo;
 import com.multi.vo.PmediVo;
@@ -39,7 +31,7 @@ import com.multi.vo.UsersVo;
 /**
  * @author noranbear
  * @date 2022. 7. 6.
- * @version 16.3
+ * @version 17.0
  * @description
  *
  *
@@ -84,17 +76,17 @@ import com.multi.vo.UsersVo;
  *  					noranbear		       medidetail 수정
  *                         				       ocraddimpl 생성
  *
- *	2022. 7. 26.						ocraddimpl에 ocrbox 실행 추가
+ *	2022. 7. 26.						 ocraddimpl에 ocrbox 실행 추가
  *
- *             			 ynr1734		    dashboard 카드 생성
+ *             			 ynr1734		     dashboard 카드 생성
  *
- *				  noranbear, qwaszx357	 ocraddimpl 기능 구현 완성
+ *				  noranbear, qwaszx357	   ocraddimpl 기능 구현 완성
  *
- *	2022. 7. 27.		noranbear		 ocraddimpl에 조건 1 추가
+ *	2022. 7. 27.		noranbear		   ocraddimpl에 조건 1 추가
  *
- *	                	najune		 		profile 업데이트 기능 
+ *	                	najune		 		 profile 업데이트 기능 
  *
- *  2022. 7. 29.		qwaszx357	  editmymedi, deletemymedi 생성
+ *  2022. 7. 29.		qwaszx357	    editmymedi, deletemymedi 생성
  *  
  *  										   dashboard 수정
  *
@@ -107,64 +99,56 @@ import com.multi.vo.UsersVo;
  *
  *                  	ynr1734             	location 생성
  *
- *	2022. 8. 5.			noranbear		mypage 생성 및 signupimpl 수정
+ *	2022. 8. 5.			noranbear		 mypage 생성 및 signupimpl 수정
  *
- *                  qwaszx357				editstop 생성
- *												          editdone 생성
+ *                  	qwaszx357				editstop 생성
+ *												editdone 생성
  *	
- *	2022. 8. 8.			noranbear    mypage, profile, update 옮기기
+ *	2022. 8. 8.			noranbear    	mypage, profile, update 옮기기
  *
- *	2022. 8. 11.								plist 수정
+ *	2022. 8. 11.								  plist 수정
  *	
- *  2022. 8. 12.		najune				addimpl ajax로 이동
+ *  2022. 8. 12.		najune				 addimpl ajax로 이동
  *  
- *                  qwaszx357				admin 생성
+ *                  	qwaszx357				  admin 생성
  *                  
- *	2022. 8. 17.						signup, signupimpl 수정
+ *	2022. 8. 17.						   signup, signupimpl 수정
  *
- *	2022. 8. 17.		noranbear		페이지의 breadcrumb 기능 구현 	
+ *	2022. 8. 17.		noranbear		  페이지의 breadcrumb 기능 구현 	
  *
  *											처방상세정보의 복약일 구현
+ *
+ *	2022. 8. 18.							api controller 이동
  *
  * ====================================================================
  */
 
 @Controller
 public class MainController {
-	
-	// 이미지 저장 경로
-	@Value("${userdir}")
-	String userdir;
-	
-	@Autowired
-	DataAPI dapi;
-  
-  	@Autowired
-	OCRBoxAPI bapi;
   
 	@Autowired
 	UsersBiz ubiz;
+	
+  	@Autowired
+  	DashBiz dbiz;
+  	
+	@Autowired
+	MymediBiz mbiz;
 	
 	@Autowired
 	PlistBiz plibiz;
 	
 	@Autowired
-	MymediBiz mbiz;
+	PmediBiz pmedibiz;
+	
+  	@Autowired
+  	AlarmBiz abiz;
 	
 	@Autowired
 	SlistBiz slibiz;
 	
 	@Autowired
 	SmediBiz smbiz;
-  
-	@Autowired
-	PmediBiz pmedibiz;
-  
-  	@Autowired
-  	DashBiz dbiz;
-  	
-  	@Autowired
-  	AlarmBiz abiz;
   	
   	int plistid = 0;	// 처방내역 id 저장
 	
@@ -180,6 +164,7 @@ public class MainController {
 		m.addAttribute("pagename2", "검색");
 		return "index";
 	}
+	
 	
 	/**
 	 * 대시보드 페이지 연결
@@ -302,6 +287,7 @@ public class MainController {
 		return "index";
 	}
 	
+	
 	/**
 	 * 로그인 페이지 연결
 	 * @return signin.html
@@ -417,42 +403,7 @@ public class MainController {
 		return "index";
 	}
 	
-	
-	
-   /**
-    * 약 디테일 페이지 연결
-    * @return medidetail.html
-    */
-   @RequestMapping("/medidetail")
-   public String mdetail(Model m, String item) {
 
-	   Object obj = dapi.dataapi(item);
-       //System.out.println("result 값 : " + obj);
-       
-       // Object를 JSONObject으로 변환
-       JSONObject jo = (JSONObject) JSONValue.parse(obj.toString());
-       //System.out.println("JSONObject로 변환 : " + jo);
-       
-       // jo에서 JSONObject으로 body 뽑아내기
-       JSONObject jo1 = new JSONObject();
-       jo1 = (JSONObject) jo.get("body");
-       //System.out.println("body 뽑아내기 : " + jo1);
-       
-       // body에서 JSONArray로 items 뽑아내기
-       JSONArray ja = new JSONArray();
-       ja = (JSONArray) jo1.get("items");
-       // System.out.println("items 뽑아내기 : " + ja);
-       
-       m.addAttribute("item", ja);
-       
-		// page breadcrumb
-		m.addAttribute("pagename", "약 상세정보");
-		m.addAttribute("pagename2", item);
-	
-       m.addAttribute("center", "medidetail");
-       return "index";
-   }
-	
    /**
 	* 내 약 리스트 페이지 연결
 	* @return mymedilist.html
@@ -463,18 +414,18 @@ public class MainController {
 	    List<MymediVo> dlist = null; 
 	    UsersVo users = null;
 	        
-	        if(session.getAttribute("signinusers") != null){
-	            users = (UsersVo) session.getAttribute("signinusers");
-		        try {
-		            list = mbiz.getusers(users.getId());
-		            m.addAttribute("mymedi", list);
-		            dlist = mbiz.getdday(users.getId());
-		            m.addAttribute("dlist", dlist);
-		        } catch (Exception e) {
-		            e.printStackTrace();
-		        }
-	        
+        if(session.getAttribute("signinusers") != null){
+            users = (UsersVo) session.getAttribute("signinusers");
+	        try {
+	            list = mbiz.getusers(users.getId());
+	            m.addAttribute("mymedi", list);
+	            dlist = mbiz.getdday(users.getId());
+	            m.addAttribute("dlist", dlist);
+	        } catch (Exception e) {
+	            e.printStackTrace();
 	        }
+        
+        }
 	    
 	    // page breadcrumb
 		m.addAttribute("pagename", "약 보관함");
@@ -497,6 +448,7 @@ public class MainController {
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
+		
 		return "redirect:mymedi";
 	}
 	
@@ -512,8 +464,10 @@ public class MainController {
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
+		
 		return "redirect:mymedi";
 	}
+	
 	
 	/**
 	 * 처방내역 페이지 연결
@@ -628,7 +582,6 @@ public class MainController {
 		return "redirect:plist";
 	}
 	
-	
 	/**
 	 * 복약 중지
 	 * @param id
@@ -641,95 +594,8 @@ public class MainController {
 		} catch (Exception e) {
 			e.printStackTrace();
 		}
+		
 		return "redirect:plist";
-	}
-	
-	/**
-	 * 메인페이지에서 ocr창으로 들어온 1. 이미지 저장, 2. ocrbox search, 
-	 * 3. 해당 내용 DB(slist, smedi)에 저장하는 함수
-	 * @param mf 이미지 파일
-	 * @param session 현재 유저 id를 가져올 HttpSession
-	 */
-	@RequestMapping("/ocraddimpl")
-	public String ocraddimpl(Model m, MfVo mf, HttpSession session) {
-		int listId = 0;				// 스캔한 약곽 정보를 담은 slistVo의 id를 담음
-		String imgname = null;		// mf의 이미지 이름을 담음
-		String uid = null;			// users에서 받을 유저 id를 담음
-		String name = null;			// 스캔한 약곽 이름을 담음
-		
-		UsersVo users = null;		// HttpSession에서 가져오는 유저 정보를 담음
-		SlistVo slist = null;		// 스캔한 약곽 정보를 slist tbl에 넣을 때 사용
-		
-
-		// [1] 약 이미지 저장
-		// 1-1. 이미지 이름을 가져온다.
-		imgname = mf.getMf().getOriginalFilename();		// mf 앞에 fake path가 붙기 때문에 필요
-
-		// 조건 1: 이미지가 존재할 때 실행
-		if(!(imgname.isEmpty())) { 
-			
-			// [2] DB에 정보 저장
-			// 2-1. (비)회원 ID를 가져온다.
-			// 유저일 때
-	        if(session.getAttribute("signinusers") != null){	
-	        	users = (UsersVo) session.getAttribute("signinusers");		// HttpSession에서 UsersVo 가져옴
-	            uid = users.getId();
-	        
-	        // 비회원일 때
-	        }else {
-	        	uid = "none0";		// DB에 저장되어 있는 비회원 아이디
-	        }
-	        
-	        slist = new SlistVo(uid, imgname);  
-	         
-			try {
-				
-				// 2-2.스캔내역 tbl에 데이터 넣는다.
-				slibiz.registerbox(slist);
-				
-				// 1-2. 이미지를 해당 경로에 저장한다.
-				Util.saveFile(mf.getMf(), userdir);
-				
-				// [3] ocrbox 스캔
-				// 3-1. ocrbox로 스캔한다.
-				Object result = bapi.boxapi(imgname);
-				
-				// 3-2. 리턴된 Object에서 Json parsing해서 약이름만 빼낸다.
-				JSONObject jo = (JSONObject) JSONValue.parse(result.toString());
-				JSONArray images = (JSONArray) jo.get("images");
-				//System.out.println("1 : " + jo1);
-				JSONObject obj = (JSONObject) images.get(0);
-				//System.out.println("2 : " + obj);
-				JSONObject title = (JSONObject) obj.get("title");
-				//System.out.println("3 : " + obj2);
-				name = (String) title.get("inferText");
-				//System.out.println("4 : " + name);
-				
-				// 3-3. 구한 약이름을 화면으로 보낸다.
-				m.addAttribute("ocrresult", "ok");
-				m.addAttribute("resultname", name);
-				
-				// 2-3. 스캔약 tbl에 데이터를 넣는다.
-				slist = slibiz.gettheone(slist);	// smedi는 sid를 필요로 하기 때문에 DB에 있는 
-													// (id를 가지고 있는)slist를 다시 꺼내온다.
-
-				// slist tbl에 해당 정보가 존재할 때
-				if(slist.getId() != 0) {		// int id 값이 없는 경우 0 리턴
-					listId = slist.getId();
-				    SmediVo smedi = new SmediVo(name, listId);
-					smbiz.register(smedi);
-				}
-				
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-		}
-		// page breadcrumb
-		m.addAttribute("pagename", "검색");
-		m.addAttribute("pagename2", "검색");
-		
-		return "index";
-
 	}
 	
 	/**
@@ -753,10 +619,9 @@ public class MainController {
 		
 		m.addAttribute("center", "location");
 		return "index";
-  }
+	}
   
-  
-  /**
+	/**
 	 * Alarm tbl에 알람 데이터를 추가
 	 * @return 처방디테일 페이지로 돌아감
 	 */
