+++ src/main/java/com/multi/mapper/DashMapper.java
@@ -18,18 +18,18 @@ import com.multi.vo.SmediVo;
  * @description
  *
  *
- * =========================================================
+ * ============================================================
  * 	    DATE			 AUTHOR				    NOTE
- * ---------------------------------------------------------
+ * ------------------------------------------------------------
  *  2022. 7. 26.		ynr1734		  		First Creation
  *  
- *	2022. 7. 29.		qwaszx357		   getsmeditoday 생성
- *											getsmeditop 생성
- *										  getsmeditopday 생성
+ *	2022. 7. 29.		qwaszx357		  getsmeditoday 생성
+ *										   getsmeditop 생성
+ *										 getsmeditopday 생성
  *
- *	2022. 8. 4.			qwaszx357	  대시보드에 출력할 스캔 정보 생성
+ *	2022. 8. 4.						  대시보드에 출력할 스캔 정보 생성
  *
- * =========================================================
+ * ============================================================
  */
 
 @Repository
