+++ src/main/java/com/multi/mapper/MymediMapper.java
@@ -20,14 +20,15 @@ import com.multi.vo.MymediVo;
  * ---------------------------------------------------------
  *  2022. 7. 11.		qwaszx357		  First Creation
  *  
- *	2022. 7. 20.		najune				get 추가 
+ *	2022. 7. 20.		najune				  get 추가 
  *
  *	2022. 7. 29.		qwaszx357		   selectdday 생성
  *
- *  2022. 8. 9.         noranbear          select5medis 생성
+ *  2022. 8. 9.         noranbear         select5medis 생성
  *
  * =========================================================
  */
+
 @Repository
 @Mapper
 public interface MymediMapper {
