+++ src/main/java/com/multi/mybatis/mymedimapper.xml
@@ -39,6 +39,5 @@ PUBLIC "-//mybatis.org/DTD Mapper 3.0//EN"
 		ORDER BY dday ASC LIMIT 5;
 	</select>
 	
-	
-	
+
 </mapper>
