+++ src/main/java/com/multi/restapi/DataAPI2.java
@@ -26,6 +26,7 @@ import org.springframework.stereotype.Component;
  *
  * =========================================================
  */
+
 @Component
 public class DataAPI2 {
