+++ src/main/java/com/multi/restapi/DataAPI3.java
@@ -28,6 +28,7 @@ import org.springframework.stereotype.Component;
  *
  * =========================================================
  */
+
 @Component
 public class DataAPI3 {
