+++ src/main/java/com/multi/vo/AlarmVo.java
@@ -16,9 +16,9 @@ import lombok.ToString;
  * ================================================================
  * 	    DATE			 AUTHOR				    NOTE
  * ----------------------------------------------------------------
- *  2022. 7. 13.		qwaszx357		  First Creation
+ *  2022. 7. 13.		qwaszx357		   First Creation
  *  
- *  2022. 8. 2.			noranbear		 오전/오후, 시간, 분 없애고 
+ *  2022. 8. 2.			noranbear		오전/오후, 시간, 분 없애고 
  *  										 time으로 통일
  *
  *	2022. 8. 4.						  complete를 button으로 바꾸고 
@@ -27,9 +27,9 @@ import lombok.ToString;
  *									updatebt 함수를 위해 id와 button 
  *									  값만 받는 Constructor 생성
  *
- *	2022. 8. 5.			qwaszx357			gage 생성
+ *	2022. 8. 5.			qwaszx357			   gage 생성
  *
- *	2022. 8. 9.		  noranbear				phos 필드 추가
+ *	2022. 8. 9.		  	noranbear		     phos 필드 추가
  *
  * ================================================================
  */
@@ -48,7 +48,7 @@ public class AlarmVo {
 	private int pid;
 	
 	private String phos;	// plist hospital - select5alarms 함수 때문에 생성
-	private int gage;	// 순응도
+	private int gage;		// 순응도
   
 	
 	// Constructor for insert
