+++ src/main/java/com/multi/vo/UsersVo.java
@@ -88,9 +88,5 @@ public class UsersVo {
 		
 	}
 	
-	
-
-	
-
 
 }
