+++ src/main/resources/templates/center.html
@@ -3,45 +3,45 @@
 <!-- Description: 메인 페이지의 기본 중간 화면 (검색창)
 
 -------------------------------------------------------------------------------
-	Date			 Author							Note
+	Date			 Author							   Note
 -------------------------------------------------------------------------------
-2022. 7. 15.	    noranbear				   First Creation
+2022. 7. 15.	    noranbear				   	   First Creation
 
-2022. 7. 16.							Bootstrap Dashboard page 탬플릿으로 바꿈
+2022. 7. 16.						    Bootstrap Dashboard page 탬플릿으로 바꿈
 
-2022. 7. 18. 								  검색 페이지 화면 구성
-                              				  main-logo path 변경
+2022. 7. 18. 								      검색 페이지 화면 구성
+                              				      main-logo path 변경
 
-2022. 7. 19.		qwaszx357				  검색 결과 출력 화면 구성
+2022. 7. 19.		qwaszx357				     검색 결과 출력 화면 구성
 
-2022. 7. 21.		noranbear				검색 결과 화면 출력 테스트 완료
+2022. 7. 21.		noranbear			      검색 결과 화면 출력 테스트 완료
 	
-2022. 7. 22.							  getdataAJAXtest.html 내용을 
-										  대입해서 테이블 내에 결과 출력하기	
+2022. 7. 22.							     getdataAJAXtest.html 내용을 
+										     대입해서 테이블 내에 결과 출력하기	
 
-					ynr1734						OCR 이미지 창 구현
+					ynr1734					      OCR 이미지 창 구현
 
-					qwaszx357					footer 지우기
+					qwaszx357					   footer 지우기
 
-2022. 7. 23.		noranbear				  검색 없음 메세지 출력
-										medidetail controller로 item 보내기 시작
+2022. 7. 23.		noranbear				     검색 없음 메세지 출력
+								       medidetail controller로 item 보내기 시작
 
-2022. 7. 25.							medidetail controller로 item 보내기 성공	
+2022. 7. 25.					       medidetail controller로 item 보내기 성공	
 
 											ocr 이미지 controller로 보내기 성공												 
 								  			
-2022. 7. 26.									ocrbox 실행 성공
+2022. 7. 26.									  ocrbox 실행 성공
 
-             	     najune						분류 검색 기능 추가	
+             	     najune						  분류 검색 기능 추가	
 
-2022. 7. 29.		qwaszx357			검색 결과 div 크기 변경, 대체 이미지 출력
+2022. 7. 29.		qwaszx357			 검색 결과 div 크기 변경, 대체 이미지 출력
 
-2022. 8. 4.			noranbear					indentation	수정				 
+2022. 8. 4.			noranbear					 indentation 수정				 
 	
-2022. 8. 17.								ocrbox 자동 검색으로 수정
+2022. 8. 17.								  ocrbox 자동 검색으로 수정
 
 
-                qwaszx357					검색 결과 테이블 수정
+                	qwaszx357					 검색 결과 테이블 수정
 			  							 
 -------------------------------------------------------------------------------
 -->
