+++ src/main/resources/templates/index.html
@@ -29,9 +29,9 @@
 												  
 											side bar에 유저 로고 추가
 												side bar 없애기
-									    메인 페이지 헤더 아이콘과 검색창 구현
+									     메인 페이지 헤더 아이콘과 검색창 구현
 									    
-2022. 7. 15							 foot 제작 및 검색창의 select list 구현
+2022. 7. 15							  foot 제작 및 검색창의 select list 구현
 
 										기존 Bootstrap 및 작업 내용 삭제
 											새 Bootstrap 연동 완료						
@@ -53,20 +53,20 @@
 												
 											    plist 링크 수정
 											  
-2022. 7. 21.		noranbear				필요없는 import 지우기
+2022. 7. 21.		noranbear				 필요없는 import 지우기
 
 2022. 7. 22.    	qwaszx357       			 footer 추가
 
-2022. 8. 4.			noranbear					하단 plugin 삭제
+2022. 8. 4.			noranbear				   하단 plugin 삭제
 
 2022. 8. 5.	      	qwaszx357				utype에 따른 navbar 변경
 		
 					noranbear				헤더 검색창 삭제 및 마이페이지 
-												링크 mypage에 연결
+											   링크 mypage에 연결
 											  
-2022. 8. 8.			 ynr1734			페이지 타이틀 변경 및 favicon 변경		
+2022. 8. 8.			 ynr1734			 페이지 타이틀 변경 및 favicon 변경		
 
-2022. 8. 12.		qwaszx357						링크 수정
+2022. 8. 12.		qwaszx357					   링크 수정
 
 ----------------------------------------------------------------------------
 -->
