+++ src/main/resources/templates/medidetail.html
@@ -25,22 +25,22 @@
 
 2022. 7. 16.		ynr1734					     화면 구성 변경
 
-2022. 7. 19			            화면 구성 변경 및 내 약 리스트에 담기 버튼 스낵바 구현
+2022. 7. 19			             화면 구성 변경 및 내 약 리스트에 담기 버튼 스낵바 구현
 
-2022. 7. 20          najune					내약리스트 버튼 구현
+2022. 7. 20          najune					  내약리스트 버튼 구현
 
 2022. 7. 22.		qwaszx357					footer 지우기
 
 2022. 7. 25.								디테일 정보 출력 기능 구현
 
-2022. 7. 26.									추가 버튼 기능 구현
+2022. 7. 26.								   추가 버튼 기능 구현
 
-2022. 7. 27.							modal 제품명, 이미지 가운데 정렬
-									추가 완료 modal 지우고 현재 modal에 출력
+2022. 7. 27.							 modal 제품명, 이미지 가운데 정렬
+									  추가 완료 modal 지우고 현재 modal에 출력
 									
-2022. 8. 4.										부작용, 관련약 출력
+2022. 8. 4.									   부작용, 관련약 출력
 
-2022. 8. 15.									주의사항에 강조효과
+2022. 8. 15.								   주의사항에 강조효과
 
 2022. 8. 17.		noranbear					화면 마무리 정리
