+++ src/main/resources/templates/mymedi.html
@@ -21,16 +21,16 @@
 -------------------------------------------------------------------------								    
 2022. 7. 16.		noranbear				  medidetail.html 복붙
 
-2022. 7. 18.		najune				  		내약리스트 화면 구현
+2022. 7. 18.		najune				  	   내약리스트 화면 구현
 
-2022. 7. 20.						  			내약리스트 기능 구현
+2022. 7. 20.						  		   내약리스트 기능 구현
 
-2022. 7. 22.		qwaszx357					head, html 지우기
+2022. 7. 22.		qwaszx357				   head, html 지우기
 
 2022. 7. 26.									image src 수정
 
-2022. 7. 29.		qwaszx357					수정, 삭제 기능 구현
-											대체 이미지 출력, modal 추가
+2022. 7. 29.								   수정, 삭제 기능 구현
+										   대체 이미지 출력, modal 추가
 
 2022. 8. 17.		noranbear				 화면 마무리 수정사항 구현										]
