+++ src/main/resources/templates/padd.html
@@ -21,7 +21,7 @@
 -------------------------------------------------------------------------								    
 2022. 7. 18			noranbear					First Creation
 
-2022. 7. 22.		najune			 	         profile 화면 구현
+2022. 7. 22.		najune			 	        profile 화면 구현
 
 2022. 7. 22.		qwaszx357					footer 지우기
