+++ src/main/resources/templates/pdetail.html
@@ -25,18 +25,18 @@
 												  
 2022. 7. 22.		qwaszx357					footer 지우기
 
-2022. 7. 25. 		najune						    기능구현
+2022. 7. 25. 		najune						   기능구현
 
-2022. 7. 30.		                  처방약 기능 구현
+2022. 7. 30.		                  			처방약 기능 구현
                 
-                noranbear					알람 추가 창 구현
+               		noranbear					알람 추가 창 구현
 
-2022. 8. 2. 								화면 로드될 때, 해당 알람내역이
-												캘린더에 뜨게 구현
+2022. 8. 2. 							  화면 로드될 때, 해당 알람내역이
+											   캘린더에 뜨게 구현
 
-2022. 8. 4.									복약 버튼 리스트 화면 구현 완료												
+2022. 8. 4.								  복약 버튼 리스트 화면 구현 완료												
 
-											calendar 불필요한 코드 정리
+										   calendar 불필요한 코드 정리
 
 											 복약 알람 화면 구현 및 
 												알람 띄우기 완료
@@ -46,7 +46,7 @@
 
 											복약 알람 없을 시 화면 구현
 
-2022. 8. 4.				qwaszx357				순응도 구현						
+2022. 8. 4.			qwaszx357				  	  순응도 구현						
 
 2022. 8. 12.								처방내역 상세 반복 없애기		
 										'복용중'일 경우에만 복용완료 버튼 클릭
