+++ src/main/resources/templates/plist.html
@@ -21,15 +21,15 @@
 -------------------------------------------------------------------------								    
 2022. 7. 18.		qwaszx357				  tables.html 가져옴
 
-2022. 7. 19.									출력 기능 구현
+2022. 7. 19.									 출력 기능 구현
 					
 2022. 7. 19.		noranbear			  카드에 링크 걸기 및 스타일 수정
 
-2022. 7. 22.		qwaszx357					footer 지우기
+2022. 7. 22.		qwaszx357					 footer 지우기
 
-													row 변경
+												   row 변경
 													
-2022. 8. 11.									   순응도 구현
+2022. 8. 11.									  순응도 구현
 
 2022. 8. 18.		noranbear				로그인 전 modal 창 띄우기
