+++ src/main/resources/templates/signin.html
@@ -23,9 +23,9 @@
 
 					qwaszx357					  code 정리
 					
-2022. 7. 16			qwaszx357					로그인 기능 구현
+2022. 7. 16										로그인 기능 구현
 
-2022. 7. 23.								   헤더 변경 및 링크 연결
+2022. 7. 23.								  헤더 변경 및 링크 연결
 					
 -------------------------------------------------------------------------
 -->
