+++ src/main/resources/templates/signup.html
@@ -23,13 +23,13 @@
 
 					qwaszx357					  code 정리
 		
-2022. 7. 16			najune						회원가입 페이지 구현	
+2022. 7. 16			najune					   회원가입 페이지 구현	
 		
-2022. 7. 18			najune						회원가입 기능 구현		
+2022. 7. 18			najune					   회원가입 기능 구현		
 
-2022. 7. 23.		qwaszx357				   헤더 변경 및 링크 연결
+2022. 7. 23.		qwaszx357				  헤더 변경 및 링크 연결
 
-2022. 7. 28.		najune				    	아이디 중복체크 기능, 
+2022. 7. 28.		najune				      아이디 중복체크 기능, 
 											 이메일 @ 유효성 체크 기능
 
 2022. 8. 5.			noranbear					성별 value 수정
